requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
# paddleocr==2.9.1        # Commented out — using Google Vision API instead
# paddlepaddle==3.0.0     # Commented out — PaddleOCR dependency
google-cloud-vision==3.8.1
opencv-python-headless==4.10.0.84
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.20
httpx==0.28.1
Pillow==11.1.0
numpy>=1.26.0
# easyocr==1.7.2          # Commented out — using Google Vision API instead
# pytesseract==0.3.13     # Commented out — using Google Vision API instead
anthropic>=0.39.0
openai>=1.50.0
thefuzz>=0.22.1
python-Levenshtein>=0.25.0
pytest-asyncio>=0.24.0
